Cartpage component is split into smaller components
