--- src/components/List.jsx
@@ -1,105 +1,0 @@
-// import { useEffect, useState } from "react";
-import useCart from "../hooks/useCart";
-import Listitem from "./Listitem";
-import useAuth from "../hooks/useAuth";
-import { useNavigate } from "react-router-dom";
-import ButtonPrimary from "./buttons/ButtonPrimary";
-
-function List() {
-  const { cart } = useCart();
-  const { auth } = useAuth();
-  const navigate = useNavigate();
-
-  function getLowestPrice(c) {
-    return c.prices.length > 0
-      ? c.prices.reduce((prev, curr) => (prev.price < curr.price ? prev : curr))
-      : null;
-  }
-
-  return (
-    <div className="max-w-[900px] w-full min-h-screen mx-auto text-center flex flex-col font-general">
-      <div className="flex flex-row justify-center mt-20 bg-[#e8e3d8] rounded">
-        <section className="container w-full p-4 mb-4 rounded md:p-10">
-          <div className="flex flex-row justify-between mb-2 align-middle">
-            <h1 className="pb-2 text-4xl text-left align-middle md:pb-4">
-              Din kurv
-            </h1>
-          </div>
-          <table className="w-[100%] select-none" id="cart">
-            <thead className="border-b-2 border-black">
-              <tr className="text-left">
-                <th className="hidden md:block"></th>
-                <th></th>
-                <th></th>
-                <th className="text-center"></th>
-              </tr>
-            </thead>
-            <tbody className="">
-              {cart.length > 0 &&
-                cart.map(product => (
-                  <Listitem key={product.product_id} product={product} />
-                ))}
-            </tbody>
-          </table>
-          {cart.length === 0 && (
-            <p className="pt-2 text-xl font-bold">Kurven er tom!</p>
-          )}
-          <div className="flex flex-col justify-between gap-2 pt-4 md:items-end">
-            <div className="flex flex-row justify-between md:gap-4 md:w-[300px]">
-              <p>Varer i alt: {cart.reduce((a, c) => c.quantity + a, 0)} stk</p>
-              <p>
-                {cart.length > 0
-                  ? cart
-                      .reduce(
-                        (a, c) => c.quantity * getLowestPrice(c).price + a,
-                        0
-                      )
-                      .toFixed(2)
-                  : "0.00"}{" "}
-                kr.
-              </p>
-            </div>
-            <div className="flex flex-row justify-between border-b-2 pb-2 border-black md:gap-4 md:w-[300px]">
-              <p>Levering</p>
-              {cart.length > 0 ? <p>59.00 kr.</p> : <p> 0.00 kr.</p>}
-            </div>
-            <div className="flex flex-row justify-between text-xl font-bold md:gap-4 md:w-[300px]">
-              <p>I alt</p>
-              <p>
-                {cart.length > 0
-                  ? (
-                      cart.reduce(
-                        (a, c) => c.quantity * getLowestPrice(c).price + a,
-                        0
-                      ) + 59
-                    ).toFixed(2)
-                  : "0.00"}{" "}
-                kr.
-              </p>
-            </div>
-            {cart.length > 0 &&
-              (auth?.accessToken ? (
-                <ButtonPrimary
-                  onClick={() => navigate("/checkout")}
-                  className={`md:w-[300px] ${
-                    cart.length === 0 && "disabled cursor-not-allowed"
-                  }`}
-                >
-                  Gå til checkud
-                </ButtonPrimary>
-              ) : (
-                <ButtonPrimary
-                  onClick={() => navigate("/login")}
-                  className={"md:w-[300px]"}
-                >
-                  Login
-                </ButtonPrimary>
-              ))}
-          </div>
-        </section>
-      </div>
-    </div>
-  );
-}
-
-export default List;
