+++ src/pages/cartpage/button/CartButton.jsx
@@ -1,0 +1,32 @@
+import { useNavigate } from "react-router";
+import ButtonFullPrimary from "../../../components/buttons/ButtonFullPrimary";
+import useAuth from "../../../hooks/useAuth";
+
+function CartButton({ cart }) {
+  const { auth } = useAuth();
+  const navigate = useNavigate();
+  return (
+    <>
+      {cart.length > 0 &&
+        (auth?.accessToken ? (
+          <ButtonFullPrimary
+            onClick={() => navigate("/checkout")}
+            className={`md:w-[300px] ${
+              cart.length === 0 && "disabled cursor-not-allowed"
+            }`}
+          >
+            Gå til checkud
+          </ButtonFullPrimary>
+        ) : (
+          <ButtonFullPrimary
+            onClick={() => navigate("/login")}
+            className={"md:w-[300px]"}
+          >
+            Login
+          </ButtonFullPrimary>
+        ))}
+    </>
+  );
+}
+
+export default CartButton;
